Read flycircuit neurons from hemibrainr Google Drive

Read flycircuit neurons from the hemibrainr google drive. The google drive must be mounted with Google Filestream or rclone. See hemibrainr_set_drive.

flycircuit_neurons(
  local = FALSE,
  folder = "light_level/flycircuit",
  cable = c("all", "primary.neurites", "arbour"),
  data = c("neuronlist", "nblast", "dps"),
  brainspace = c("JRCFIB2018F", "JRCFIB2018Fraw", "FCWB")
)

lm_lhns(
  local = FALSE,
  folder = "light_level/lhns",
  data = c("neuronlist", "nblast", "dps"),
  cable = c("lhns", "lhins", "lines"),
  brainspace = c("JRCFIB2018F", "JRCFIB2018Fraw", "FCWB")
)

Arguments

local
FALSE or path. By default (FALSE) data is read from options()$Drive_hemibrain_data), but the user can specify an alternative path.

folder
A subfolder on the hemibrain team drive or your local data folder containing the data object to read.

cable
the type of cable we want to read. "all" indicates full neurons.

data
the type of data to read, i.e. neurons, an NBLAST matrix or a dotprops object.

brainspace
A template brain space for neurons loaded by lm_lhns. Defaults to JRCFIB2018F.

Value

a nat::neuronlistfh object for 'light-level' neurons in the given brainspace.

See also

Examples

# \donttest{
if (FALSE) {

# All flycircuit neurons
fw = flycircuit_neurons()

# All LHNs from Frechter et al. 2019, eLife
lhns = lm_lhns()

}# }