Read precomputed flywire data from the hemibrainr Google Drive

Read precomputed data available on the hemibrain Google Team Drive. (see hemibrainr_set_drive) and (see hemibrainr_rclone). This includes body IDs for all flywire neurons ((flywire_ids)), and user contributions towards their creation (flywire_contributions), as well as Flywire related NBLAST scores retrieved using hemibrain_nblast.

flywire_meta(local = FALSE, folder = "flywire_neurons/", sql = TRUE, ...)

flywire_contributions(
  local = FALSE,
  folder = "flywire_neurons/",
  sql = TRUE,
  ...
)

flywire_failed(local = FALSE, folder = "flywire_neurons/", sql = TRUE, ...)

flywire_ids(local = FALSE, folder = "flywire_neurons/", sql = TRUE, ...)

Arguments

local
FALSE or path. By default (FALSE) data is read from options()$Drive_hemibrain_data), but the user can specify an alternative path.

folder
A subfolder on the hemibrain team drive or your local data folder containing the data object to read.

sql
logical. Whether not to read the desired data from an SQL database. This can save you from having to load a lot of information into memory, and it works well with dplyr pipes. If FALSE the relevant .csv is read from the connected google drive (see hemibrainr_set_drive) and (see hemibrainr_rclone) and loaded into memory, which can take some time.

...
if sql=TRUE, methods passed to dplyr::tbl.

Value

a data.frame. Depending on which synapse function was called, it can contain the columns:

"flywire.xyz" - coordinates of a point in the neuron in flywire voxel space. XYZ, separated by a semicolon.

"flywire.id" - the unique ID associated with this flywire neuron. This ID changes every time a neuron is, even slightly, modified. So it is an unstable identifier. This is why flywire.xyz is sometimes used.

"fw.x" - the x coordinate of a point in the flywire neuron, in flywire voxel space..

"fw.y" - the y coordinate of a point in the flywire neuron, in flywire voxel space..

"fw.z" - the z coordinate of a point in the flywire neuron, in flywire voxel space..

"user_name" - the name of the user who made the number of edits given in this row.

"edits" - the number of edits (merges, splits, etc.) made by a user for the given flywire.id.

"proportion" - the proportion of total edits for this neuron, that the given user made.

"dataset" - the dataset this neuron is from, i.e. flywire.

See also

Examples

# \donttest{
if (FALSE) {

# All flywire IDs for neurons that have a split precomputed
fw.ids = flywire_ids()

# For these flywire IDs, their meya data:
fw.meta = flywire_meta()

# For flywire IDs, which users contributed what:
fw.edits = flywire_contributions()

}# }