Find the antennal lobe glomerulus for hemibrain body ids

Find the antennal lobe glomerulus for hemibrain body ids

glomerulus(bodyids, exclude.multi = FALSE)

Arguments

bodyids
Vector of bodyids or a search string passed to neuprint_ids. See details.

exclude.multi
When TRUE multiglomerular neurons will be returned as NA.

Value

A character vector of glomeruli named with the bodyids. If no glomerulus exists (or the bodyid was NA on input) NA_character_ is returned.

Details

Note that this is based on the same hemibrain metadata used by class2ids. These were curated by Marta Costa, Elizabeth Marin, Philipp Schlegel and Greg Jefferis in the Drosophila Connectomics Group / MRC LMB in Cambridge.

bodyids may contain duplicates and NA values (in which case NA_character_ will be returned).

See also

Examples

if (FALSE) {
table(glomerulus(class2ids('PN')), useNA='ifany')
table(glomerulus(class2ids('PN'), exclude.multi = TRUE), useNA='ifany')
}