The hemibrain mesh

A 3D mesh of the overall neuropil structures in the hemibrain volume as a hxsurf object. Created by Philipp Schlegel from the union of individual ROIs in FlyEM's hemibrain project (see details). Mesh available both in raw voxels and in microns.

hemibrain.surf

hemibrain_microns.surf

Format

An object of class hxsurf (inherits from list) of length 4.

An object of class hxsurf (inherits from list) of length 4.

Source

https://neuprint.janelia.org/

Details

The hemibrain mesh is watertight and contains no internal structure or non-manifold vertices. It is the right thing to use if you want to see if things are inside the hemibrain neuropil. Philipp Schlegel used the following steps to construct the mesh:

Taking all the individual ROIs

Voxelizing them

Combining the voxels into a single image stack

Filling the gaps in between (binary operations + some manual fixing in Fiji)

Using a marching cubes to generate a mesh from the stack

Note that hemibrain coordinate system does not have the anterior-posterior axis aligned with the Z axis.

See also

Examples

# \donttest{
library(nat)
if (FALSE) {
plot3d(hemibrain.surf, alpha=.3, col='grey')

# can also convert to rgl mesh3d object for more control
hemibrain <- as.mesh3d(hemibrain.surf)
rgl::wire3d(hemibrain)
rgl::shade3d(hemibrain, alpha=.3)
rgl::shade3d(hemibrain, alpha=.3, col='grey')
}
# }