Manage hemibrain-FAFB neuron matches

We can match neurons in the hemibrain data with FAFB neurons (hemibrain->FAFB, hemibrain tab) and FAFB neurons have been matched with their hemibrain counterparts (FAFB->hemibrain, fafb tab). These matches are recorded on a Google Sheet on our hemibrain Google Team Drive, to which you will need access through an authenticated account to view and use this function. We can use these function to manipulate the Google Sheet. This include adding new IDs for matching (hemibrain_matching_add), transfer matches between the sheet's hemibrain and FAFB tabs (hemibrain_matching_transfers) and add pre-made matches (hemibrain_add_matches).

hemibrain_add_made_matches(
  df,
  direction = c("both", "hemibrain-FAFB", "FAFB-hemibrain"),
  ...
)

hemibrain_matching_add(
  ids,
  sheet = c("hemibrain", "FAFB"),
  User = "flyconnectome",
  selected_file = "1OSlDtnR3B1LiB5cwI5x5Ql6LkZd8JOS5bBr-HTi0pOw",
  ...
)

hemibrain_matching_transfers(
  selected_file = "1OSlDtnR3B1LiB5cwI5x5Ql6LkZd8JOS5bBr-HTi0pOw"
)

Arguments

df
a data.frame of pre-made matches that the user wants to transfer onto the Google Sheet. This will erase any extant matches for the specified neurons. This data frame must have columns: bodyid (i.e. the hemibrain neuron), skid (i.e. the FAFB neuron) and quality (i.e. match quality). Matches have three levels of 'quality', largely dependent on the degree of manual tracing for FAFB neurons - good (could be the same cell), medium (same cell type) and poor (could be the same or similar cell type). @param ids either hemibrain bodyids or FAFB skids to add to a Google Sheet. You will want to add IDs where they do not already exist, so that you can use, for example, hemibrain_matching. @param sheet the tab to which to add your new information. You are either adding to information related ot hemibrain neurons, or FAFB neurons.

direction
the match direction, i.e. hemibrain->FAFB (hemibrain tab) or FAFB->hemibrain (fafb tab). Defaults to updating both.

...
arguments passed to methods for, for example, neuprintr::neuprint_get_meta and elmr::fafb_get_meta.

User
your initials, so updates can be linked to you. There is a limited number of users, each of whom have been assigned a number of neurons to match up. In order to add yourself as a user, simply open this Google Sheet in your browser and add your initials to neurons of your choosing on the rightmost column 'Users'.

selected_file
Specifies which Google Sheet to use. Unless you are using a personal Google Sheet, this should be "1OSlDtnR3B1LiB5cwI5x5Ql6LkZd8JOS5bBr-HTi0pOw".

Value

NULL. Updates the master Google sheet.

Details

Currently, information is recorded in a Google Sheet

See also

Examples

# \donttest{
if (FALSE) {

# Add a mising FAFB projection neuron, so we can match it later:
hemibrain_matching_add(ids = "16", sheet = "fafb", User = "ASB")

# Match interactively!
fafb_matching(ids="16", overwrite = TRUE)

}# }