Manage hemibrain-FAFB neuron matches

We can match neurons in the hemibrain data with FAFB neurons (hemibrain->FAFB, hemibrain tab) and FAFB neurons have been matched with their hemibrain counterparts (FAFB->hemibrain, fafb tab). These matches are recorded on a Google Sheet on our hemibrain Google Team Drive, to which you will need access through an authenticated account to view and use this function. We can use these function to manipulate the Google Sheet. This include adding new IDs for matching (hemibrain_matching_add), transfer matches between the sheet's hemibrain and FAFB tabs (hemibrain_matching_transfers) and add pre-made matches (hemibrain_add_matches).The function lywire_matching_rewrite updates a Google sheet with correct flywire IDs and stable locations in flywire space, based on CATMAID skids.

flywire_matching_rewrite(
  flywire.ids = names(flywire_neurons()),
  selected_file = options()$hemibrainr_matching_gsheet,
  ...
)

hemibrain_add_made_matches(
  df,
  direction = c("both", "hemibrain-FAFB", "FAFB-hemibrain"),
  User = "flyconnectome",
  ...
)

hemibrain_matching_add(
  ids,
  dataset = c("hemibrain", "FAFB", "flywire"),
  User = dataset,
  selected_file = options()$hemibrainr_matching_gsheet,
  ...
)

hemibrain_matching_transfers(
  selected_file = options()$hemibrainr_matching_gsheet
)

fafb_matching_rewrite(
  selected_file = options()$hemibrainr_matching_gsheet,
  ...
)

hemibrain_matching_rewrite(
  selected_file = options()$hemibrainr_matching_gsheet,
  ...
)

Arguments

flywire.ids
flywire IDs to add to Google sheet if not already present.

selected_file
Specifies which Google Sheet to use. Unless you are using a personal Google Sheet, this should be options()$hemibrainr_matching_gsheet.

...
arguments passed to methods for, for example, neuprintr::neuprint_get_meta and elmr::fafb_get_meta.

df
a data.frame of pre-made matches that the user wants to transfer onto the Google Sheet. This will erase any extant matches for the specified neurons. This data frame must have columns: bodyid (i.e. the hemibrain neuron), skid (i.e. the FAFB neuron) and quality (i.e. match quality). Matches have three levels of 'quality', largely dependent on the degree of manual tracing for FAFB neurons - good (could be the same cell), medium (same cell type) and poor (could be the same or similar cell type).

direction
the match direction, i.e. hemibrain->FAFB (hemibrain tab) or FAFB->hemibrain (fafb tab). Defaults to updating both.

User
your initials, so updates can be linked to you. There is a limited number of users, each of whom have been assigned a number of neurons to match up. In order to add yourself as a user, simply open this Google Sheet in your browser and add your initials to neurons of your choosing on the rightmost column 'Users'.

ids
either hemibrain bodyids or FAFB skids to add to a Google Sheet. You will want to add IDs where they do not already exist, so that you can use, for example, hemibrain_matching.

dataset
the tab to which to add your new information. You are either adding to information related to hemibrain neurons, or FAFB neurons.

Value

NULL. Updates the master Google sheet.

Details

Currently, information is recorded in a Google Sheet

See also

Examples

# \donttest{
if (FALSE) {

# Add a mising FAFB projection neuron, so we can match it later:
hemibrain_matching_add(ids = "16", sheet = "FAFB", User = "ASB")

}# }