Adjust and correct hemibrain neurons somas.

Currently provides three "modes": correct single neurons, correct cell body fibers, and correct from google sheet. Correct singles will allow you to manually correct each neuron in a provided list of bodyids. If requested, DBSCAN can be used to try to predict potential soma positions against already correctly identified somas, based on the provided google sheet. Correct cell body fibers(cbf) asks for a cbf to be inputted based on those labelled in the google sheet. DBSCAN is then used to cluster the already annotated soma positions on the google sheet within the cbf. This will allow you to identify correct soma positions, and then manually fix incorrect somas. for each neuron with an incorrect soma, a potential new soma will be suggested based on the DBSCAN clustering result. Correct google sheet is useful when screening large numbers of neurons. Once a large group of neurons has been split into morphological clusters, using NBLAST for example, this mode will request a cluster id and apply DBSCAN to somas within the cluster. You can then quickly screen to identify correct soma clusters, and label incorrect somas to be corrected separately.

hemibrain_adjust_saved_somas(
  bodyids = NULL,
  c = NULL,
  brain = NULL,
  selected_file = "1YjkVjokXL4p4Q6BR-rGGGKWecXU370D1YMc1mgUYr8E",
  db = NULL,
  plot_sample = TRUE,
  eps = NULL,
  minPts = NULL,
  neurons_from_gsheet = TRUE,
  for_Imaan = FALSE
)

Arguments

bodyids
Optional, unless using neurons method. Otherwise, list of neuron bodyids

c
Optional. For use when correcting Cell body fibers. If not input, function will ask you to enter one later

brain
Optional. by default will use the hemibrain surface. Other neuropil surfaces can be provided here however

selected_file
Identifier for gsheet you wish to read from and write to. By default, Curated_splitpoints soma sheet

db
Optional. if provided, local directory to read neurons from.

plot_sample
logical. TRUE by default, choose if you wish to plot a random subset of neurons within a DBSCAN cluster

eps
the distance in nanometres used by DBSCAN to form clusters. 1500 by default

minPts
The minimum number of points needed to form a cluster using DBSCAN. 5 by default

neurons_from_gsheet
logical, TRUE by default. If true, will collect neurons based on bodyids in the google sheet, otherwise, will search for them based on CBF data in neuprint.

for_Imaan
extra little bit for Imaan... FALSE by default

Value

Updates Google Sheet with soma information