Meshes for Antennal Lobe (AL) glomeruli

3D meshes of 51 olfactory + 7 thermosensory antennal lobe glomeruli for the the hemibrain volume. They are represented as a hxsurf object, in which 'regions' are defined, one for each glomerulus. Meshes are available both in raw voxels and in microns. You can see which glomeruli are truncated, and by roughly how much by examining. The glomeruli are made from PN dendrites. hemibrain_glomeruli_summary.

hemibrain_al.surf

hemibrain_al_microns.surf

hemibrain_glomeruli_summary

odour_scenes

flywire_al.surf

Format

An object of class hxsurf of length 4.

An object of class hxsurf of length 4.

An object of class data.frame with 58 rows and 11 columns.

An object of class data.frame with 83 rows and 7 columns.

An object of class hxsurf of length 4.

Source

https://neuprint.janelia.org/

Value

hemibrain_glomeruli_summary is a data.frame with the columns:

"expected_cit"the citation for the number of expected RNs,

"expected_RN_female_1h".

"expected_RN_female_SD".

"missing".

"odour_scenes"the general 'odour scene(s)' which this glomerulus may help signal (Mansourian 2015, Bates 2020)

"key_ligand"the ligand that excites the cognate ALLRN or receptor the most, based on pooled data from multiple studies (Münch 2016)).

"valence"the presumed valence of this odour channel (Badel 2016).

Details

note that hemibrain coordinate system has the anterior-posterior axis aligned with the Y axis (rather than the Z axis, which is more commonly observed).

These meshes were generated by Sri Jagannathan and Tomke Stuerner based on the location of PN dendritic synapses.

See also

Examples

# \donttest{
nvertices(hemibrain_al.surf)

#> [1] 162970

materials(hemibrain_al.surf)

#>      name id     col
#> D       D  1 #C70E7B
#> DA1   DA1  2 #DA2F7D
#> DA2   DA2  3 #EE5080
#> DA3   DA3  4 #E16A88
#> DA4l DA4l  5 #8471A0
#> DA4m DA4m  6 #2778B8
#> DC1   DC1  7 #1188C5
#> DC2   DC2  8 #30A0CB
#> DC3   DC3  9 #4FB8D0
#> DC4   DC4 10 #84A794
#> DL1   DL1 11 #BE904E
#> DL2d DL2d 12 #EF7F15
#> DL2v DL2v 13 #F19530
#> DL3   DL3 14 #F2AC4A
#> DL4   DL4 15 #CDB455
#> DL5   DL5 16 #73AB4B
#> DM1   DM1 17 #19A141
#> DM2   DM2 18 #25AE2F
#> DM3   DM3 19 #5AC419
#> DM4   DM4 20 #8FDA04
#> DM5   DM5 21 #9AAD10
#> DM6   DM6 22 #A6801C
#> DP1l DP1l 23 #B66E36
#> DP1m DP1m 24 #CF9E71
#> V       V 25 #E9CEAD
#> VA1d VA1d 26 #E6C6BB
#> VA1v VA1v 27 #CD95A7
#> VA2   VA2 28 #B56493
#> VA3   VA3 29 #C57EAB
#> VA4   VA4 30 #DBA5CB
#> VA5   VA5 31 #EFC3DA
#> VA6   VA6 32 #EFA78C
#> VA7l VA7l 33 #EF8B3E
#> VA7m VA7m 34 #EF851D
#> VC1   VC1 35 #F19C37
#> VC2   VC2 36 #F3B252
#> VC3   VC3 37 #E69756
#> VC4   VC4 38 #D46850
#> VC5   VC5 39 #C23A4B
#> VL1   VL1 40 #D66949
#> VL2a VL2a 41 #EC9948
#> VL2p VL2p 42 #F9C047
#> VM1   VM1 43 #F5D347
#> VM2   VM2 44 #F0E646
#> VM3   VM3 45 #C7E947
#> VM4   VM4 46 #80E04A
#> VM5d VM5d 47 #3BD74C
#> VM5v VM5v 48 #279C50
#> VM6   VM6 49 #1C5A53
#> VM7d VM7d 50 #1E2255
#> VM7v VM7v 51 #6F2E4E
#> VP1d VP1d 52 #BF3B48
#> VP1l VP1l 53 #EA3C39
#> VP1m VP1m 54 #E13020
#> VP2   VP2 55 #D92307
#> VP3   VP3 56 #A5472E
#> VP4   VP4 57 #607E6E
#> VP5   VP5 58 #1BB6AF

if (FALSE) {
plot3d(hemibrain_al.surf, alpha=.3)

# can also plot single glomeruli
plot3d(subset(hemibrain_al.surf,"DM1"))

}
# }