Return consistent cell body fiber annotation for neurons

Return consistent cell body fiber annotation for neurons

hemibrain_cbf(x, ..., meta = FALSE)

Arguments

x
A vector of bodyids, a query or a data.frame compatible with neuprint_ids

...
Additional arguments passed to neuprint_get_meta

meta
Whether to return just the name of the cell body fibers (when meta=FALSE, the default) or a data frame with additional metadata, as returned by neuprint_get_meta.

Value

a vector of new style cell body fiber names or a data.frame (as returned by neuprint_get_meta with an extra column cbf containing the preferred cell body fiber names).

Details

New style cell body fiber names are of the form ADL01 and match the regular expression '([AP][DV][ML][0-9]+)'. As of hemibrain v1.0.1 these seem to be variably located in the cellBodyFiber and instance fields and in some cases are missing from both.

See also

Examples

# \donttest{
# these are all one lineage but some missing tract info
table(hemibrain_cbf("adPN"), useNA = 'ifany')

#> 
#> AVM01 AVM03  <NA> 
#>    78     1    33 

# Do these come from two different hemilineages?
table(hemibrain_cbf("lPN"))

#> 
#> AVM02 
#>    44 

# }