Get useful metrics for each neuron's axon and dendrite

Calculates/retrieves useful metrics for the axon and dendrites of a neuron, and its split.

hemibrain_compartment_metrics(
  x,
  resample = 10,
  delta = 62.5,
  locality = FALSE,
  ...
)

Arguments

x
a nat::neuronlist or nat::neuron object. It is assumed that this neuron has been read in by neuprintr::neuprint_read_neurons or possibly catmaid::read.neurons.catmaid.

resample
stepsize to which to resample neurons. If set to NULL, neurons are not resampled.

delta
the distance at which a synapse might occur

locality
if TRUE hemibrain_overlap_locality is called and the overlap locality score will be returned in addition to other metrics.

...
Additional arguments passed to methods or eventually to nat::nlapply

Value

a data.frame of metrics broken down by putative axon and dendrite, and their overlap_locality scores

See also

Examples

# \donttest{

# Choose neurons
## In this case some antennal lobe local neurons
al.local.neurons = c("1702323386", "2068966051", "2069311379", "1702305987", "5812996027",
"1702336197", "1793744512", "1976565858", "2007578510", "2101339904",
"5813003258", "2069647778", "1947192569", "1883788812", "1916485259",
"1887177026", "2101348562", "2132375072", "2256863785", "5813002313",
"5813054716", "5813018847", "5813055448", "1763037543", "2101391269",
"1794037618", "5813018729", "2013333009")

# use a smaller group in examples
al.local.neurons=al.local.neurons[1:5]

# Get neurons
neurons = neuprint_read_neurons(al.local.neurons)

# Convert to microns
neurons = hemibrainr::scale_neurons(neurons)

# Split
neurons = hemibrain_flow_centrality(neurons)

#> Warning: 2 neurons cropped, split likely to be inaccurate for: 2068966051, 2069311379

#> Warning: 3 neurons have no soma tagged, split could be inaccurate for: 1702323386, 2068966051, 2069311379

# Calculate overlap and other metrics
hcm = hemibrain_compartment_metrics(neurons,
resample = NULL,
just.leaves = TRUE, # just look how much the leaves of skeletons overlap, quicker (default)
delta = 5)

# }