Download all automatically split neurons from the hemibrain project

Download all the automatically split neurons from the hemibrain project, from Google Drive. The neuronlist is saved as a nat::neuronlistfh object so certain neurons may be read from it without loading the entire, large neuronlist into memory. In order to do this, you need to add the following folder to your Google Drive: https://drive.google.com/drive/folders/14UPg7CvHDtvzNjvNgAULYnxZ018Xgf5H?usp=sharing. Contact us if you do not have, but would like, permission.

hemibrain_download_neurons(savedir = TRUE, overwrite = FALSE, ...)

Arguments

savedir
where to save the .rds and meta data files for the nat::neuronlistfh object

overwrite
whether or not to overwrite the existing saved files, within savedir.

...
Additional arguments passed to nlapply when iterating over neurons to download.

See also

Examples

# \donttest{
if (FALSE) {
# Parallelise
numCores <- parallel::detectCores()
doMC::registerDoMC(numCores/2)
message("Using ", numCores/2, " cores")

# Download neurons
hemibrain_download_neurons(.parallel = TRUE, overwrite = FALSE)

# Get specific neurons
neurons = hemibrain_read_neurons("1702323386", savedir = TRUE)
}# }