Extract synapse location and labels, and edgelists from a neuron/neuronlist

Extract a single data frame describing synapse/connection types, partners, locations and position on a neuron's axon/dendrite. You can either get all synapses returned or all unitary connections to a neuron's partners returned. Broken down by axon/dendrite (Label), and pre/postsynapses or pre/postsynaptic partners. Note that hemibrain_extract_compartment_edgelist will only return connections between neurons given in the argument x.

hemibrain_extract_synapses(x, prepost = c("BOTH", "PRE", "POST"), ...)

hemibrain_extract_connections(x, prepost = c("BOTH", "PRE", "POST"), ...)

hemibrain_extract_compartment_edgelist(x, ...)

Arguments

x
a nat::neuronlist or nat::neuron object. It is assumed that this neuron has been read in by neuprintr::neuprint_read_neurons or possibly catmaid::read.neurons.catmaid.

prepost
whether to get presynapses, postsynapses or both

...
Additional arguments passed to methods or eventually to nat::nlapply

Value

a data.frame

See also

Examples

if (FALSE) {
# Choose bodyids
al.local.neurons =
 c("1702323386", "2068966051", "2069311379", "1702305987", "5812996027",
 "1702336197", "1793744512", "1976565858", "2007578510", "2101339904",
 "5813003258", "2069647778", "1947192569", "1883788812", "1916485259",
"1887177026", "2101348562", "2132375072", "2256863785", "5813002313",
 "5813054716", "5813018847", "5813055448", "1763037543", "2101391269",
 "1794037618", "5813018729", "2013333009")

# Read in these neurons
neurons = neuprintr::neuprint_read_neurons(al.local.neurons)

# Re-root
neurons.flow = flow_centrality(neurons, polypre = TRUE,
  mode = "centrifugal", split = "distance")

# Let's check that this worked
syns = hemibrain_extract_synapses(neurons.flow)

# Get the edgelist by compartment
elist = hemibrain_extract_compartment_edgelist(neurons.flow)

# See result
library(nat)
plot3d_split(neuron.flow)
points3d(xyzmatrix(subset(syns,prepost==1)), col = "cyan")
points3d(xyzmatrix(subset(syns,prepost==0)), col = "red")
}