Read precomputed information from the hemibrain Google Drive

Read precomputed data available on the hemibrain Google Team Drive. This includes all synapses, neuron-neuron connections and an edgelist for all hemibrain neurons, broken down by axon and dendrite assignments. NBLAST matrices for all neurons against all neurons in the data set are also available, including ones broken down by neuron compartment.

hemibrain_neuron_bodyids(savedir = TRUE, local = FALSE)

hemibrain_elist(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_synapses(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_connections(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_nblast(
  nblast = c("all", "primary.neurites", "primary.dendrites", "axons", "dendrites",
    "spines", "tracts", "arbour", "simplified"),
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

flycircuit_neurons(
  savedir = TRUE,
  local = FALSE,
  folder = "hemibrain_neurons/light_level/flycircuit",
  cable = c("all", "primary.neurites", "arbour"),
  data = c("neuronlist", "nblast", "dps")
)

Arguments

savedir
directory in which neuron data has been deposited. If TRUE your default save directory is used, which is stored as: getOption("Gdrive_hemibrain_data")

local
logical, whether to try to read locally saved neurons (by default at: getOption("hemibrain_data")) or neurons from Google Drive (getOption("Gdrive_hemibrain_data")).

neuron.split
read saved neurons split in which way? Folder names indicative of arguments passed to flow_centrality. and hemibrain_flow_centrality

nblast
the NBLAST matrix you would like to retrieve, e.g. "arbours" gives you a normalised all by all NBLAST matrix of all branching arbour.

folder
A subfolder containing the data object to read.

cable
the type of cable we want to read. "all" indicates full neurons.

data
the type of data to read, i.e. neurons, an NBLAST matrix or a dotprops object.

Value

a data.frame or character vector

See also

Examples

# \donttest{
if (FALSE) {

#  All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_edgelist()

}# }