Manually verified neuron matches between the hemibrain and FAFB dataset

Manually verified neuron matches between the hemibrain and FAFB dataset, as well as some light level data from the package natverse-compatible lhns. The Fly Connectome team at the University of Cambridge has manually made cell-to-cell matches based on an initial NBLAST calculation, and given matches a qualitative score. There can be no match (none), a poor match (poor) an okay match (okay) or an exact match (good). As a rule of thumb, a poor match could be a neuron from a very similar same cell type or a highly untraced neuron that may be the correct cell type. An okay match should be a neuron that looks to be from the same morphological cell type but there may be some discrepancies in its arbour. An exact match is a neuron that corresponds well between FAFB and the hemibrain data. the Google Sheet or using hemibrain_matches if they have an authenticated email.

hemibrain_matched

lm_matched

Format

An object of class data.frame with 99155 rows and 16 columns.

An object of class data.frame with 27878 rows and 11 columns.

Value

a data.frame where each row is a neuron, either from the hemibrain or FAFB data sets. Each row gives you its matching neuron in the other data set. These matches have been manually assigned using fafb_matching, hemibrain_matching and LR_matching. If you use this information make sure you credit it appropriately. Contact us if unsure:

"cell.type" - the neuPrint designated 'type' for the neuron. If dataset is not "hemibrain", then this is based on the hemibrain match.

"cell" - the unique cell, which is just cell.type#number.

"cellBodyFiber" - the cell body fiber to which this neuron belongs

"ItoLee_Hemilineage" - the hemilineage to which this neuron belongs. Seer hemibrain_hemilineages.

"match" - the ID of the manual match from the other data set. If dataset=="hemibrain" then this is a flywire.id that can be found in flywire_neurons.If "CATMAID" or "flywire" then it is a hemibrain body ID.

"quality" - the matcher makers qualitative assessment of how good this match is.

"FAFB.hemisphere.match" - the flywire coordinates of a neuron on the opposite hemisphere, which a match maker has designated as this id's cognate.

"FAFB.hemisphere.match.quality" - the quality of this match.

"LM.match" - indicates a light level neuron that is a match for id. This neuron will be in flycircuit_neurons() or other light level data.

"LM.match.quality" - the quality of this match.

"dataset" - the data set to which id belongs. Either "hemibrain", or one of the two FAFB repositories, "CATMAID" or "flywire".

"priority" - whether FAFB->hemibrain matches ("FAFB") or hemibrain->FAFB matches ("hemibrain") were used in order to ascribe cell type names to FAFB neurons. In both cases, cell type names are attached to hemibrain bodyids, and propagated to their FAFB matches.