Retrieve matched up neurons between the hemibrain and FAFB

Many neurons in the hemibrain data have been matched one to one with FAFB neurons (hemibrain->FAFB) and FAFB neurons have been matched with their hemibrain counterparts (FAFB->hemibrain). These matches have been done by the Flyconnectome Group at the University of Cambridge, and are recorded on a Google Sheet on our hemibrain Google Team Drive, to which you will need access through an authenticated account to view and use this function. Matches have three levels of 'quality', largely dependent on the degree of manual tracing for FAFB neurons - good (could be the same cell), medium (same cell type) and poor (could be the same or similar cell type).

hemibrain_matches(priority = c("FAFB", "hemibrain"))

Arguments

priority
whether to use FAFB->hemibrain matches (FAFB) or hemibrain->FAFB matches (hemibrain) in order to ascribe cell type names to FAFB neurons. In both cases, cell type names are attached to hemibrain bodyids, and propagated to their FAFB matches.

Value

a data.frame which includes neuron's ID (either its CATMAID skeleton ID or neuprint body ID), the data set from which it comes, its putative cell type and connectivity type, and its match in the other dataset.

Details

Currently, the Google Sheet is set up with limited number of users, each of whom have been assigned a number of neurons to match up. In order to add yourself as a user, simply open this Google Sheet in your browser and add your initials to neurons of your choosing on the rightmost column 'Users'.

See also

Examples

# \donttest{
if (FALSE) {

# Get matches
matched = hemibrain.matches()

}# }