Axon-dendrite split points calculated for all hemibrain neurons

Axon-dendrite split points calculated for the whole hemibrain dataset, hemibrain:v1.0.1, using a variety of parameters. The file used to generate this can be found in this package, as data-raw/hemibrain_split_all_neurons.R. The arguments given to flow_centrality are in the name of the data frame. Note that, because the proportion of recovered synapses is unevenly distributed across brain regions, after compartment detection, sometimes the dendrite was chosen as the neurite closest to a neuron's primary branch point (data frames with 'distance' in their name), see reference.

hemibrain_all_splitpoints

hemibrain_splitpoints_polypre_centrifugal_distance

hemibrain_splitpoints_pre_centrifugal_distance

hemibrain_splitpoints_polypre_centrifugal_synapses

Format

An object of class data.frame with 793749 rows and 18 columns.

An object of class data.frame with 794165 rows and 7 columns.

An object of class data.frame with 505094 rows and 6 columns.

An object of class data.frame with 794164 rows and 7 columns.

Source

https://www.biorxiv.org/content/10.1101/2020.01.19.911453v1

References

Bates, A. S., P. Schlegel, R. J. V. Roberts, N. Drummond, I. F. M. Tamimi, R. Turnbull, X. Zhao, et al. 2020. “Complete Connectomic Reconstruction of Olfactory Projection Neurons in the Fly Brain.” bioRxiv. doi:10.1101/2020.01.19.911453.

See also

Examples

# \donttest{
head(hemibrain_splitpoints_polypre_centrifugal_distance)

#>    X.1    bodyid position                point       X       Y       Z
#> 1    2 202916528        2                 root 20075.3 12279.3 12198.2
#> 2  299 202916528      299 primary.branch.point 22493.3 19472.5 17848.7
#> 3  545 202916528      545         axon.primary 31914.0 24874.0 12060.0
#> 4  311 202916528      311     dendrite.primary 22930.9 20018.9 17567.6
#> 5  499 202916528      499          axon.start1 31066.0 25818.0 14524.0
#> 6 1240 202916528     1240          axon.start2 31162.0 25802.0 14588.0

if (FALSE) {
View(hemibrain_splitpoints_polypre_centrifugal_distance)
}
# }