Axon-dendrite split points calculated for all hemibrain neurons

Axon-dendrite split points calculated for the whole hemibrain dataset, hemibrain:v1.0.1. The file used to generate this can be found in this package, as data-raw/hemibrain_split_all_neurons.R. Because the proportion of recovered synapses is unevenly distributed across brain regions, after compartment detection, the dendrite was chosen as the neurite closest to a neuron's primary branch point.

hemibrain_splitpoints_distance

Format

An object of class data.frame with 505094 rows and 6 columns.

Source

https://www.biorxiv.org/content/10.1101/2020.01.19.911453v1

References

Bates, A. S., P. Schlegel, R. J. V. Roberts, N. Drummond, I. F. M. Tamimi, R. Turnbull, X. Zhao, et al. 2020. “Complete Connectomic Reconstruction of Olfactory Projection Neurons in the Fly Brain.” bioRxiv. https://doi.org/10.1101/2020.01.19.911453.

See also

Examples

# \donttest{
head(hemibrain_splitpoints_distance)

#>      bodyid position                 point       X       Y       Z
#> 1 202916528        2                  root 20075.3 12279.3 12198.2
#> 2 202916528      326  primary.branch.point 23704.7 20727.3 17166.7
#> 3 202916528      545      axon.primary.545 31914.0 24874.0 12060.0
#> 4 202916528     2008 dendrite.primary.2008 23402.0 20554.0 17564.0
#> 5 202916528      499           axon.start1 31066.0 25818.0 14524.0
#> 6 202916528     1240           axon.start2 31162.0 25802.0 14588.0

if (FALSE) {
View(hemibrain_splitpoints_distance)
}
# }