Plot chosen neurons and see all neurons in that type

Plot your selected neuron (rightmost pane) and then see all of the neurons in the hemibrain project that belong to the types you have plotted (left pane).

hemibrain_type_plot(
  bodyids = NULL,
  someneuronlist = NULL,
  brain = hemibrainr::hemibrain.surf,
  print = TRUE,
  meta = NULL
)

hemibrain_multi3d(..., someneuronlist = hemibrain_neurons())

Arguments

bodyids
bodyids for the neurons you want to plot. If someneuronlist is null then hemibrain_neurons will be used to load a neuron database.

someneuronlist
a neuronlist or neuron object to plot. If bodyids is NULL then bodyids = names(someneuronlist).

brain
a mesh3d or hxsurf object, such as a template brain, to plot alongside neurons of interest.

print
logical, whether or not to print the name of the cell types being plotted and their Body IDs.

meta
a data.frame with a bodyid and a type field, such as one pulled by neuprint_get_meta. If NULL then meta will be acquired as neuprint_get_meta(bodyids).

...
vector lists of bodyids, each of which will be plotted in a different set of colours.

Value

Plots coloured neuron(s)

See also

Examples

# \donttest{
if (FALSE) {

# Choose neurons
ids = c("605063393", "573708714", "485387561", "5813083717")

# Plot
hemibrain_type_plot(ids)

# Plot multi
open3d()
a = c("546511447", "573333903", "577546843", "573333284", "573683455",
 "609941228", "604368022")
b = c("573670049", "579576294", "727838223", "696432315")
hemibrain_multi3d(a, b)

}# }