Split a neuron using stored 'splitpoints'

Split a neuron using a data.frame of stored split points.

hemibrain_use_splitpoints(x, df, knn = FALSE, ...)

Arguments

x
a nat::neuronlist or nat::neuron object. It is assumed that this neuron has been read in by neuprintr::neuprint_read_neurons or possibly catmaid::read.neurons.catmaid.

df
a data.frame of splitpoints from running flow_centrality, as produced by hemibrain_splitpoints.

knn
logical, whether or not to find corresponding points between df and k using a nearest neighbour search in 3D space (TRUE) or just use the given point IDs in df (i.e. if neurons have not been resampled or their skeletons otherwise modified).

...
methods sent to nat::nlapply

Value

a neuronlist

See also

Examples

# \donttest{

# Choose neurons
## These neurons are some 'tough' examples from the hemibrain:v1.0.1
### They will split differently depending on the parameters you use.
tough = c("5813056323", "579912201", "5813015982", "973765182", "885788485",
"915451074", "5813032740", "1006854683", "5813013913", "5813020138",
"853726809", "916828438", "5813078494", "420956527", "486116439",
"573329873", "5813010494", "5813040095", "514396940", "665747387",
"793702856", "451644891", "482002701", "391631218", "390948259",
"390948580", "452677169", "511262901", "422311625", "451987038"
)

# Get neurons
neurons = neuprint_read_neurons(tough)

#> Warning: Dropping given bodyids that could not be read from https://neuprint.janelia.org : 5813056323

# Now make sure the neurons have a soma marked
## Some hemibrain neurons do not, as the soma was chopped off
neurons.checked = hemibrain_skeleton_check(neurons, meshes = hemibrain.surf)

#> Re-rooting 3 neurons without a soma

#> Removing synapses at somas and along primary neurite for 29 neurons

# Split neuron
## These are the recommended parameters for hemibrain neurons
neurons.flow = flow_centrality(neurons.checked, polypre = TRUE,
mode = "centrifugal",
split = "distance")

#> Warning: At structural_properties.c:745 :Couldn't reach some vertices

#> Warning: At structural_properties.c:745 :Couldn't reach some vertices

# Save the results
splitpoints = hemibrain_splitpoints(neurons.flow)

# Re-use the results
neurons.flow.2 = hemibrain_use_splitpoints(neurons, splitpoints, knn = FALSE)

#> Warning: ‘>=’ not meaningful for factors

#> Error in Summary.factor(structure(6L, .Label = c("1245", "1246", "273", "283", "284", "370", "371", "382", "384", "392", "396", "400", "402", "517", "1", "1069", "1072", "1165", "1389", "1565", "1568", "1578", "1961", "2310", "2321", "236", "2619", "2623", "2738", "2753", "2755", "2758", "2761", "2860", "2895", "2906", "2945", "3259", "3263", "3272", "347", "3582", "3589", "3593", "368", "3757", "3759", "3766", "3768", "3895", "3925", "4036", "4051", "4056", "4061", "4084", "4087", "4104", "4110", "1011", "1033", "1058", "1059", "1324", "1326", "183", "2334", "2371", "2380", "2392", "2400", "2559", "2571", "3187", "3212", "3214", "3217", "3219", "3220", "3224", "436", "506", "518", "723", "941", "101", "2524", "2566", "278", "279", "297", "1166", "1282", "1298", "1337", "1366", "1514", "1545", "2567", "2579", "2589", "2626", "2630", "2633", "295", "3102", "3119", "3133", "3136", "3137", "3149", "561", "579", "585", "73", "82", "840", "843", "10021", "10025", "11616", "11919", "12141", "12143", "12265", "2982", "3054", "3056", "3059", "3064", "3714", "3729", "4696", "4752", "4765", "4786", "7385", "7396", "7505", "7509", "7567", "7677", "7786", "7906", "7939", "7946", "7967", "8033", "8053", "9490", "9494", "9497", "9500", "9503", "9516", "9531", "9536", "9540", "9543", "9546", "9557", "9560", "9562", "9579", "9585", "9591", "9798", "9807", "9842", "9887", "9893", "9990", "135", "140", "2133", "2144", "2302", "2322", "2324", "2326", "2336", "2339", "2482", "2561", "2716", "2762", "2765", "2781", "2819", "3065", "3080", "3081", "3101", "3107", "3110", "3116", "375", "376", "100", "104", "175", "1821", "1824", "237", "260", "261", "2934", "3445", "3482", "3733", "3883", "3912", "3916", "3918", "3962", "4096", "2", "296", "300", "3648", "3658", "3666", "3838", "3945", "3957", "4215", "4242", "4244", "4246", "4248", "4250", "4254", "4259", "4262", "4264", "4265", "4270", "4272", "632", "695", "740", "741", "746", "1042", "1076", "1082", "1085", "1087", "330", "5309", "5347", "5375", "5379", "5384", "5422", "5482", "5495", "550", "5506", "5510", "553", "5610", "5626", "5629", "5632", "6", "6794", "6801", "6887", "6890", "6899", "6903", "6910", "1053", "1055", "1284", "1289", "1572", "1812", "1815", "1817", "1822", "574", "861", "862", "885", "887", "892", "902", "904", "907", "1205", "1207", "1208", "1247", "1249", "1705", "2231", "2232", "2239", "2241", "2244", "2248", "2252", "2259", "2265", "2271", "2274", "2276", "1257", "1263", "1264", "1383", "1395", "3223", "3235", "3238", "3438", "3442", "3443", "3447", "3449", "3456", "3459", "3466", "3469", "3481", "3489", "3493", "3623", "3626", "3640", "2318", "2412", "3023", "3031", "3411", "3597", "3608", "3609", "3614", "3620", "3629", "3633", "3635", "3638", "3656", "3662", "3702", "3706", "3736", "2142", "2148", "2174", "2176", "2178", "244", "287", "435", "640", "641", "653", "675", "699", "704", "866", "871", "12", "1557", "1584", "1596", "1611", "1616", "234", "369", "3854", "3905", "4153", "4204", "4241", "4273", "4277", "4286", "4288", "4293", "4298", "4299", "1086", "1089", "3478", "3529", "3664", "3668", "3678", "3681", "3703", "4095", "4351", "5112", "5141", "5158", "5161", "5166", "5169", "5173", "5179", "5183", "5186", "5189", "5192", "5194", "932", "2883", "2884", "2918", "2920", "2925", "2933", "2935", "2938", "2948", "587", "734", "735", "742", "87", "922", "1943", "1955", "1956", "2061", "2635", "3333", "3336", "3346", "3348", "3351", "3373", "3419", "3545", "3643", "3655", "3659", "3661", "1945", "1947", "3216", "3329", "3332", "3377", "3412", "3572", "3580", "3583", "3587", "3592", "3612", "3645", "1516", "1519", "1522", "1527", "1644", "1662", "239", "240", "253", "2584", "2588", "2591", "2594", "2595", "2603", "97", "2180", "2184", "2185", "2299", "2306", "2890", "3129", "3357", "3358", "3360", "1553", "1562", "1625", "1696", "1856", "1857", "3091", "3127", "3243", "3405", "3407", "4353", "4357", "4360", "9396", "9397", "9435", "9457", "9461", "9464", "9466", "9470", "9477", "1484", "1491", "1492", "1684", "1685", "1691", "1693", "1708", "1713", "1717", "1725", "1733", "2652", "2656", "2990", "3004", "1180", "1195", "1251", "1259", "2347", "2349", "2378", "2442", "2668", "2671", "3622", "5797", "5810", "5813", "5818", "5822", "985", "986", "1396", "209", "313", "325", "5846", "5850", "5853", "5862", "5864", "5867", "5875", "1459", "1460", "1465", "3573", "4232", "4252", "4258", "4261", "894", "897", "1971", "1981", "1982", "2236", "2269", "2437", "2469", "314", "315", "577", "1221", "1488", "1494", "1508", "1511", "1521", "1523", "1526", "1528", "1530", "1532", "1534", "1567", "247", "4", "488", "5", "538"), class = "factor"), na.rm = FALSE): ‘max’ not meaningful for factors

if (FALSE) {
# Plot the split to check it
nat::nopen3d()
nlscan_split(neurons.flow2, WithConnectors = TRUE)

}# }