Read precomputed hemibrain data from the hemibrainr Google Drive

Read precomputed data available on the hemibrain Google Team Drive. (see hemibrainr_set_drive) and (see hemibrainr_rclone). This includes body IDs for all hemibrain neurons ((hemibrain_neuron_bodyids)), all synapses (hemibrain_synapses), neuron-neuron connections (hemibrain_connections) and an edgelist (hemibrain_elist) for all hemibrain neurons, broken down by axon and dendrite assignments. Hemibrain related NBLASTs retrieved using hemibrain_nblast.

hemibrain_neuron_bodyids(
  local = FALSE,
  folder = "hemibrain_neurons/",
  sql = FALSE,
  ...
)

hemibrain_elist(local = FALSE, folder = "hemibrain_neurons/", sql = TRUE, ...)

hemibrain_meta(local = FALSE, folder = "hemibrain_neurons/", sql = TRUE, ...)

hemibrain_synapses(
  local = FALSE,
  folder = "hemibrain_neurons/",
  sql = TRUE,
  ...
)

hemibrain_connections(
  local = FALSE,
  folder = "hemibrain_neurons/",
  sql = TRUE,
  ...
)

Arguments

local
FALSE or path. By default (FALSE) data is read from options()$Drive_hemibrain_data), but the user can specify an alternative path.

folder
A subfolder on the hemibrain team drive or your local data folder containing the data object to read.

sql
logical. Whether not to read the desired data from an SQL database. This can save you from having to load a lot of information into memory, and it works well with dplyr pipes. If FALSE the relevant .csv is read from the connected google drive (see hemibrainr_set_drive) and (see hemibrainr_rclone) and loaded into memory, which can take some time.

...
if sql=TRUE, methods passed to dplyr::tbl.

Value

a data.frame. Depending on which synapse function was called, it can contain the columns:

"treenode_id" - the position of the node in the SWC-style table found at neuron$d, where the neuron is the skeleton for bodyid.

"connector_id" - the unique ID for a pre/post synapse, as read from neuPrint. If this is not given, you are looking at a connection not a synapse. In this case count should be given, which shows the number of synapses in this connection.

"prepost" - whether the given synapse is a pre-synapse (0, output synapse) or postsynapse (1, input synapse). Alternatively, if a connection is given, whether this connection is presynaptic to bodyid (0, bodyid is target) or postsynaptic (1, bodyid is source).

"x" - x coordinate for the root point.

"y" - y coordinate for the root point.

"z" - z coordinate for the root point.

"confidence" - FlyEM's confidence level. The lower the score, the more likely this synapse is an artefact.

"bodyid" - The neuPrint neuron/body related to the synapse/connection given in each row.

"partner" - The neuron connecting to bodyid by the given synapse/connection.

"pre" - The body ID for the presynaptic (source) neuron.

"partner" - The body ID for the presynaptic (target) neuron.

"Label" - The compartment of the bodyid neuron on which the synapse is placed / which receives/makes the given connection. See ?standardise.

"partner.Label" - The compartment of the partner neuron on which the synapse is placed / which receives/makes the given connection.

"count" - The number of synapses that make the given connection. Sometimes referred to as 'weight'.

"norm" - The normalised synapse weight. count is divided by the total number of inputs that the target neuron's (post) compartment (Label) has. I.e. this normalisation is by total_inputs onto a dendrite or axon, not the whole neuron.

See also

Examples

# \donttest{
if (FALSE) {

# All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# For these body IDs, all synapses:
syns = hemibrain_synapses()

# For these body IDs, input and output connections:
conns = hemibrain_connections()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_elist()

}# }