Set data source for precomputed information

One can read precomputed data available on the hemibrain Google Team Drive, and other supported team drives. This includes all synapses, neuron-neuron connections and an edgelist for all hemibrain neurons, broken down by axon and dendrite assignments. NBLAST matrices for all neurons against all neurons in the data set are also available, including ones broken down by neuron compartment.

hemibrainr_set_drive(
  Gdrive = "hemibrain",
  path = "/Volumes/GoogleDrive/Shared drives/"
)

hemibrainr_team_drive()

hemibrainr_rclone(
  Gdrive = "hemibrainr",
  path = paste0(getwd(), "hemibrainr_rclone/")
)

hemibrainr_rclone_unmount(path = paste0(getwd(), "hemibrainr_rclone/"))

Arguments

Gdrive
name of the mounted google drive. Can just be the name of a team drive to be found in the standard location: "/Volumes/GoogleDrive/Shared\ drives/".

path
path to your 'hemibrainr google drive', either a mounted remote or a local copy. For example on a Mac, Google filestream mounts it at: /Volumes/GoogleDrive/Shared\ drives/. With rclone sheet you can mount it wherever you like, by default hemibrainr expects it at: paste0(getwd(),"hemibrainr_rclone/")

Value

Sets options Gdrive_hemibrain_data

Details

In order to use this package to its fullest, you need to get hemibrainr to read large amounts of data stored on a google drive. To do this, you have three options:

1. Mount your google drive usin Google filestream 2. Mount your google drive using rclone 3. Download the google drive and save locally

The best option is to use google filestream. By default, this is what hemibrainr expects. However, you need a Google Workspace account (formerly G-Suite), which is a paid-for service. Without this, the best option is to use hrefhttps://rclone.org/drive/rclone. For detailed instructions on how to confiure rclone, please see your article in this package: Reading hemibrainr data from google drive. You can also download to an external hard drive and use that.

We have two Google team drives available for you to use, which contain similar data. One ("hemibrain") is for internal use by the Drosphila Connctomcis Group. The other one ("hemibrainr") is shared with those who would like access. Contact us by email to request access.

See also

Examples

# \donttest{
if (FALSE) {
# The default drive is named 'hemibrain'
# The Wilson lab at HMS uses a drive called 'hemibrainr'.
# You could also set up your own drive with saved data.
hemibrainr_set_drive("hemibrainr")

# Ah but what it you do not have the dirve loaded
# With google filestream say, as a network drive?
# In this case, you can use the free-service rclone instead
# However, you will need to install and configure it:https://rclone.org/drive/
hemibrainr_rclone(Gdrive = "hemibrainr")
### Mounts in your working directory

# All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_elist()

# Unmount rclone drive
hemibrainr_rclone_unmount()

}# }