Meta information of certain neuron classes

For olfactory projection neurons (PNS), antennal lobe local neurons (ALLN), non-MB third order olfactory neurons (TONs), lobula connection visual neurons (LC), Kenyon cells (KC) and MB output neurons (MBONs). Collated by the Flyconnectome group.

pn.info

mbon.info

ton.info

lc.info

alln.info

kc.info

cent.info

dn.info

Format

An object of class data.frame with 347 rows and 57 columns.

An object of class data.frame with 71 rows and 60 columns.

An object of class data.frame with 2383 rows and 48 columns.

An object of class data.frame with 2129 rows and 45 columns.

An object of class data.frame with 197 rows and 48 columns.

An object of class data.frame with 1927 rows and 45 columns.

An object of class data.frame with 24 rows and 45 columns.

An object of class data.frame with 346 rows and 33 columns.

Details

"bodyid" - unique identifier for a single hemibrain neuron.

"pre" - the number of presynapses (outputs) a neuron contains. Each of this is polyadic.

"post" - the number of postsynapses (inputs) to the neuron.

"upstream" - the number of incoming connections to a neuron.

"downstream" - the number of outgoing connections from a neuron.

"soma" - neuron size in voxels.

"sid" - which brain hemisphere contains the neuron's soma.

"voxels" - whether the neuron has a soma in the hemibrain volume.

"name" - the name of this neuron.

"connectivity.type" - the neuPrint designated 'type' for the neuron. A subset of neurons within a cell_type that share similar connectivity. A connectivity type is distinguished from a cell_type by an ending _letter, defined using CBLAST (Scheffer et al. 2020).

"cell_type" - neurons of a shared morphology that take the same cell body fibre tract and come from the same hemilineage (Bates 2019).

"cellBodyFiber" - The cell body fiber for a neuron, as read from neuPrint and annotated by a team under Kei Ito.

"putative.classic.transmitter" - putative neurotransmitter based on what neurons in the hemilineage in question has been shown to express, out of acetylcholine, GABA, glutamate.

"ito_lee_hemilineage" - the hemilineage that we reckon this cell_type belongs to, based on expert review of light level data from the K. Ito and T. Lee groups (Yu 2013, Ito 2013). See hemibrain_hemilineages.

"hartenstein_hemilineage" - the hemilineage that we reckon this cell_type belongs to, based on expert review of light level data from the V. Hartenstein Lee group (Wong 2013, Lovick 2013). See hemibrain_hemilineages.

"FAFB.match" - the ID of the manual match from the FAFB data set. ID indicates a neuron reconstructed in FAFBv14 CATMAID. Many of these neurons will be available through Virtual Fly Brain.

"FAFB.match.quality" - the matcher makers qualitative assessment of how good this match is: a poor match could be a neuron from a very similar same cell_type or a highly untraced neuron that may be the correct cell_type. An okay match should be a neuron that looks to be from the same morphological cell_type but there may be some discrepancies in its arbour. A good match is a neuron that corresponds well between FAFB and the hemibrain data.

"layer" - probabilistic mean path length from neuron from ALRNs, depends on connection strengths.

"layer.ct"- the mean layer for cell_type, rounded to the nearest whole number.

"axon_outputs"- number of outgoing connections from the neuron's predicted axon.

"dend_outputs"- number of outgoing connections from the neuron's predicted dendrite.

"axon_inputs"- number of incoming connections from the neuron's predicted axon.

"dend_inputs"- number of incoming connections from the neuron's predicted dendrite.

"total_length"- total cable_length of the neuron in micrometers.

"axon_length"- total axon cable_length of the neuron in micrometers.

"dend_length"- total dendrite cable_length of the neuron in micrometers.

"pd_length"- total cable_length of the primary dendrite 'linker' between axon and dendrite.

"segregation_index"- a quantification of how polarised a neuron is, in terms of its segregation of inputs onto its predicted dendrite and outputs onto its axon, where 0 is no-polarisation and 1 is totally polarised (Schneider-Mizell 2016).

"notes"- other notes from annotators.

"expected_cit"a citation that describes this glomerulus,

"expected_RN_female_1h".

"expected_RN_female_SD".

"missing".

"odour_scenes"the general 'odour scene(s)' which this glomerulus may help signal (Mansourian 2015, Bates 2020)

"key_ligand"the ligand that excites the cognate ALLRN or receptor the most, based on pooled data from multiple studies (Münch 2016)).

"valence"the presumed valence of this odour channel (Badel 2016).