Download light-level lineage clone data from Ito et al. 2013 or Yu et al. 2013

Get registered light-level clonal data of central brain, secondary developmental lineages. Most neurons in the fly brain hail from one of these lineages. You can find their meta information in hemibrain_hemilineages.

lineage_clones(
  x = NULL,
  local = FALSE,
  brain = c("FAFB14", "FlyWire", "JRCFIB2018Fraw", "JRCFIB2018F", "FAFB", "JFRC2",
    "JFRC2013", "JRC2018F", "FCWB"),
  dataset = c("Ito", "Lee")
)

lineage_converter(
  x,
  from = c("ItoLee_Hemilineage", "cellBodyFiber", "ItoLee_Lineage",
    "Hartenstein_Lineage", "Hartenstein_Hemilineage"),
  to = "Hartenstein_Hemilineage"
)

Arguments

x
a specific lineage, e.g. ALlv1. If NULL, the default, all lineages are fetched

local
FALSE or path. By default (FALSE) data is read from options()$Drive_hemibrain_data), but the user can specify an alternative path.

brain
the brainspace in which hemibrain neurons have been registered. Defaults to raw voxel space for the FlyWire project.

dataset
whether to get clonal data from Ito et al. 2013 "Ito" or Yu et al. 2013 "Lee".

from
the name system in which you can find x. One of: cellBodyFiber, ItoLee_Lineage, ItoLee_Hemilineage, Hartenstein_Lineage, Hartenstein_Hemilineage

to
the lineage name system you want to move to, can be multiple of: cellBodyFiber, ItoLee_Lineage, ItoLee_Hemilineage, Hartenstein_Lineage, Hartenstein_Hemilineage

Value

A neuronlist object containing point clouds for neuron lineage clones. In the meta-data, it might be useful for some users to note that you will get:

"Hartenstein_Lineage" The Hartenstein lineage name based on expert review of light level data from the V. Hartenstein group (Wong 2013, Lovick 2013). See hemibrain_hemilineages.

"ItoLee_Lineage" Lineage name, data from the K. Ito and T. Lee groups (Yu 2013, Ito 2013). See hemibrain_hemilineages

"side" The hemisphere of the brain onto which this clone is registered

References

Ito M, Masuda N, Shinomiya K, Endo K, Ito K. Systematic analysis of neural projections reveals clonal composition of the Drosophila brain. Curr Biol. 2013 Apr 22;23(8):644-55. doi: 10.1016/j.cub.2013.03.015. Epub 2013 Mar 28. PMID: 23541729.

Yu HH, Awasaki T, Schroeder MD, Long F, Yang JS, He Y, Ding P, Kao JC, Wu GY, Peng H, Myers G, Lee T. Clonal development and organization of the adult Drosophila central brain. Curr Biol. 2013 Apr 22;23(8):633-43. doi: 10.1016/j.cub.2013.02.057. Epub 2013 Mar 28. PMID: 23541733; PMCID: PMC3637848.

See also

Examples

# \donttest{
if (FALSE) {
# Loads all processed flywire neurons as a neuronlistfh object
lee.lins = lineage_clones(brain = "FAFB14", dataset = "Lee")
ito.lins = lineage_clones(brain = "FAFB14", dataset = "Ito")

# Set the scene
library(elmr)
nopen3d()
plot3d(FAFB)

# Plot WEDd1
lee.lins.wedd1 = subset(lee.lins, ItoLee_Lineage == "WEDd1")
plot3d(xyzmatrix(lee.lins.wedd1),
add = TRUE, col = hemibrain_bright_colors["green"])

# Plot WEDd1 from the other clonal data set
ito.lins.wedd1 = subset(ito.lins, ItoLee_Lineage == "WEDd1")
plot3d(xyzmatrix(ito.lins.wedd1),
add = TRUE, col = hemibrain_bright_colors["darkgreen"])

# Get flywire neurons
fw.neurons = flywire_neurons()
fw.neurons.wedd1 = subset(fw.neurons, ItoLee_Hemilineage == "WEDd1")
plot3d(fw.neurons.wedd1,
col = hemibrain_bright_colors["marine"], soma = 4000)

# What are the equivalent hemibrain neurons?
cbf.hl = subset(hemibrain_hemilineages, ItoLee_Lineage == "WEDd1")
hb.neurons = hemibrain_neurons(brain="FAFB14")
hb.neurons.wedd1 = subset(hb.neurons,cellBodyFiber%in%cbf.hl$cellBodyFiber)
plot3d(hb.neurons.wedd1,
col = hemibrain_bright_colors["orange"], soma = 4000)

}# }