Plot neurons split up by flow centrality

Plot neurons split up by flow centrality

plot3d_split(
  someneuronlist,
  col = c("#1BB6AF", "#EF7C12", "#C70E7B", "#8FDA04", "#4D4D4D", "#FC6882"),
  splitnode = FALSE,
  WithConnectors = TRUE,
  WithNodes = F,
  soma = NULL,
  soma.alpha = 1,
  highflow = FALSE,
  lwd = 1,
  radius = NULL,
  already_selected = NULL,
  check.template = FALSE,
  ...
)

nlscan_split(
  someneuronlist,
  already_selected = NULL,
  col = c("#1BB6AF", "#EF7C12", "#C70E7B", "#8FDA04", "#4D4D4D", "#FC6882"),
  WithConnectors = TRUE,
  WithNodes = FALSE,
  soma = NULL,
  radius = NULL,
  highflow = FALSE,
  Verbose = TRUE,
  Wait = TRUE,
  sleep = 0.1,
  selected_file = NULL,
  selected_col = "#fadadd",
  yaml = TRUE,
  brain = FALSE,
  ...
)

plot3d_somas(
  someneuronlist,
  col = hemibrain_bright_colors["green"],
  radius = 300,
  soma.alpha = 1,
  volume = NULL,
  invert = FALSE
)

Arguments

someneuronlist
a neuronlist or neuron object that has been modified by flow.centrality.

col
colours of sections. Defaults to orange or axons, green for primary dendrite, blue for dendrites and pink for nodes with no flow.

splitnode
if TRUE, a magenta sphere is placed at the location of the axon-dendrite split. Possible a putative action potential initiation site?

WithConnectors
whether to plot the anatomical location of pre (red) and post (cyan) synapses.

WithNodes
whether to plot branch points

soma
whether to plot a soma, and what the radius should be. If NULL, an appropriate value is guessed.

soma.alpha
numeric, alpha transparency value for plotting the soma. Passed to rgl::spheres3d

highflow
whether to plot the nodes of highest (with in one standard deviation less than maximum) flow centrality (pink points)

lwd
Line width (default 1)

radius
For connectors and axon-dendrite split node (default 1). If NULL, an appropriate value is guessed.

already_selected
a vector of IDs for neurons that are already considered selected. If NULL then no neurons are considered selected to start with.

check.template
check which template space someneuronlist is in, in order to set default plotting settings.

...
additional arguments passed to methods.

Verbose
logical indicating that info about each selected neuron should be printed (default TRUE)

Wait
logical indicating that there should be a pause between each displayed neuron

sleep
time to pause between each displayed neuron when Wait=TRUE

selected_file
an optional path to a yaml file that already contains a selection

selected_col
the color in which selected neurons (such as those specified in selected_file) should be plotted

yaml
a logical indicating that selections should be saved to disk in (human-readable) yaml rather than (machine-readable) rda format

brain
a template brain to plot. FALSE results in no brain plotted.

volume
a mesh3d or hxsurf object. Only somas outside this volume will be plotted.

invert
logical, if TRUE only somas outside volume will be plotted, if FALSE, only those inside.

Value

Plots coloured neuron(s)

See also