Download a large set of well-traced skeletonised neurons from FlyWire

Get a large number of skeletonised neurons from FlyWire. The neuronlist is saved as a nat::neuronlistfh object so certain neurons may be read from it without loading the entire, large neuronlist into memory. You will need access to the hemibrain Google Team Drive and have it mounted with Google filestream.The function flywire_neurons_update can be used to update the available data. If you want to flag flywire neurons that should be added to the Google drive, without doing this yourself, you can use flywire_request.

flywire_neurons(
  local = FALSE,
  brain = c("FlyWire", "JRCFIB2018Fraw", "JRCFIB2018F", "FAFB", "FAFB14", "JFRC2",
    "JFRC2013", "JRC2018F", "FCWB"),
  mirror = FALSE,
  ...
)

flywire_neurons_update(
  x,
  brain = c("FlyWire", "JRCFIB2018F", "FAFB", "JFRC2", "JFRC2013", "JRC2018F", "FCWB"),
  mirror = FALSE,
  ...
)

flywire_nblast_update(
  x = NULL,
  nblast = c("hemibrain-flywire", "flywire", "flywire-mirror"),
  ...
)

flywire_request(
  request,
  selected_file = options()$flywire_flagged_gsheet,
  sheet = "flywire",
  ...
)

Arguments

local
FALSE or path. By default (FALSE) data is read from options()$Drive_hemibrain_data), but the user can specify an alternative path.

brain
the brainspace in which hemibrain neurons have been registered. Defaults to raw voxel space for the FlyWire project.

mirror
logical, whether or not to read neurons that have been mirrored (i.e. flipped to the 'other' brain hemisphere).

...
Additional arguments passed to nat::nlapply.and/or fafbseg::skeletor.

x
flywire IDs to update, for the saved Google drive neuronlistfh objects called with flywire_neurons.

nblast
which flywire NBLAST to update on Google drive.

request
a neuronlist, matrix of x,y,z position or flywire ID to add to a Google sheet that records flywire positions flagged to be processed into neuron skeletons that can be called by flywire_neurons.

selected_file
the Google sheet onto which to add new flywire coordinate. I.e. Google sheet.

sheet
the tab onto which to add your requests.

Value

A neuronlist object containing flywire skeletons. In the meta-data, it might be useful for some users to note that you will get:

"flywire.id" The ID given to the corresponding volumetric body in flywire. These are used to do things like fetch volumes and are the input to the skeletor function. However, they are highly volatile and change a lot with active tracing.

"flywire.xyz" The voxel coordinate of a single point in this neuron, usually a cell body fiber tract position. This is a more accurate way of keeping tract of neuron as it will always correspond to the same 'neuron' even though its related flywire.id will change with merge/split events in flywire.

"hemilineage" An estimated hemilineage identity from both of two naming systems, Ito et al. 2013 and Wong et al. 2013

"side" An estimate as to the 'side', i.e. brain hemisphere, in which the neuron lies

"skid" The 'skeleton ID' of this neuron's match in CATMAID for FAFBv14

"FAFB.xyz" The coordinates in nanometres of a point in the neuron, in FAFBv14 space

"hemibrain_match" The bodyid of an estimated hemibrain match

See also

Examples

# \donttest{
if (FALSE) {
fw.neurons = flywire_neurons()
}# }