Retrieve or go to a specific google sheet describing flywire neurons.

These functions can help you either go to or read the right tab (flywire_tracing_sheet) on a large google sheet, or read the whole sheet and all of its tabs ((flywire_tracing_sheets), note, memoised for 30 minutes). You can also see if a flywire neuron of interest in in a google sheet (flywire_in(fw.meta = flywire_tracing_sheets())), and see if a flywire neuron of interest is among the data stored on the hemibrainr drive, i.e. (flywire_in(fw.meta = flywire_meta())).

flywire_tracing_sheet(
  regex,
  open = FALSE,
  selected_sheet = options()$flywire_lineages_gsheet,
  Verbose = TRUE,
  ...
)

flywire_tracing_sheets(
  regex = NULL,
  selected_sheet = options()$flywire_lineages_gsheet
)

flywire_in(
  query,
  query.type = c("flywire.id", "flywire.xyz", "flywire.svid"),
  regex = NULL,
  fw.meta = flywire_meta(),
  cloudvolume.url = NULL,
  Verbose = TRUE
)

Arguments

regex
character vector, tabs (i.e. work sheets) on the google sheet to query/read. This works with regex, so you only need to give the name partially. If set to NULL for flywire_tracing_sheets, the whole google sheet is read and all tabs are combined using plyr::rbind.fill.

open
logial, if TRUE the relevant google sheet tab is opened in your browser. Else, a data.frame of the tab is returned.

selected_sheet
character vector indicating the a flywire tracing google sheet. This defaults to the master 'lineage' tracing sheet used by the Drosphila Connectomics Group to store annotations on flywire neurons in different developmental lineages. Note: You may not have access.

Verbose
logical, if TRUE then hemibrainr communicates what it has found.

...
methods passed to googlesheets4 functions. .

query
a vector of data indicating flywire neurons. These can be a vector of flywire XYZ positions (readable by nat::xyzmatrix), flywire root IDs or flywire supervoxel IDs. These are used to get the most up to date root IDs from FlyWire, which are then matched to the results from flywire_tracing_sheets.

query.type
wtherh the qurey is a vector of xyz positions, flywire supervoxel IDs or flywire IDs ("flywire.xyz"/"flywire.svid"/"flywire.id").

fw.meta
a data.frame of meta data on flywire neurons, e.g. as produced by flywire_tracing_sheets.or read by flywire_meta.

cloudvolume.url
URL for CloudVolume to fetch segmentation image data. The default value of NULL choose

Value

a data.frame with columns from the right tab, chosen using regex, from the given google sheet(s) selected_sheet, specified using the argument chosen.columns.

Details

flywire tracing google sheets should have the columns: flywire.xyz (a single cardinal position for each neuron in raw FlyWire voxel space), flywire.svid (the supervoxel ID for this position) and flywire.id (the flywire root ID for the neuron that can be found at this position). Additional, informative columns are encouraged. A sheet in this format can be updated using flywire_ids_update. This updates the 'unstable' flywire.id column based on the stable flywire.xyz column, and also provides stable flywire.svid column if it is empty.

See also

Examples

# \donttest{
if (FALSE) {

# Quick look at what we have on the google sheet
sheetlist=googlesheets4::sheet_properties(
"1QyuHFdqz705OSxXNynD9moIsLvZGjjBjylx5sGZP2Yg")
View(sheetlist)

# Go to specific tab of lineage google sheet
flywire_tracing_sheet("AOTUv1_v", open = TRUE)
flywire_tracing_sheet("CREa1_v.*_r", open = TRUE)

# Get this tab as a data frame
flywire.lineage.sheet = flywire_tracing_sheet("CREa1_v.*_l")

# Get the whole sheet as a data frame
flywire.lineage.sheet = flywire_tracing_sheets()
## Note, the memoisation lasts for 30 minutes.
## I.e. this takes a long time to run but is saved for
## 30 mins on yuor computer. So if the function is called again
## with the same arguments, the sheets are not read again.

# See if we have a certain flywire neuron in our sheets
in.lin = flywire_in(query = "111437.5,21121.25,2661",
query.type = "flywire.xyz", Verbose = TRUE,
fw.meta = flywire_tracing_sheets())

# And in the meta data for neurons on the google drive
in.meta = flywire_in(query = "111437.5,21121.25,2661",
query.type = "flywire.xyz", Verbose = TRUE)

}# }