Download all automatically split neurons from the hemibrain project

Download all the automatically split neurons from the hemibrain project, from Google Drive. The neuronlist is saved as a nat::neuronlistfh object so certain neurons may be read from it without loading the entire, large neuronlist into memory. Note, you do not need to do this if you have your Google Drive mounted on your computer and you have access to the Flyconnectome Repository 2020lhfib. In which case, you should be able to read neurons directly from the drive whiled you have an Internet connection using hemibrain_read_neurons(savedir=TRUE). Neurons downloaded into your working directory.

In order to do download all neurons locally, you need to add the following folder to your Google Drive: https://drive.google.com/drive/folders/14UPg7CvHDtvzNjvNgAULYnxZ018Xgf5H?usp=sharing. Contact us if you do not have, but would like, permission. If this takes to much time, you can also download the relevant Google Drive folder manually. To do so, open this link: https://drive.google.com/open?id=1px6o2R_heFLRCtTF4Q2SvJNR9CWFhud0. Remember to unzip all files.

hemibrain_download_neurons(overwrite = FALSE, ...)

Arguments

overwrite
whether or not to overwrite the existing saved files, within savedir.

...
Additional arguments passed to nlapply when iterating over neurons to download.

See also

Examples

# \donttest{
if (FALSE) {
# Parallelise
numCores <- parallel::detectCores()
doMC::registerDoMC(numCores/2)
message("Using ", numCores/2, " cores")

# Download neurons
hemibrain_download_neurons(.parallel = TRUE, overwrite = FALSE)

# Get specific neurons
neurons = hemibrain_read_neurons("1702323386", savedir = TRUE, local = TRUE)
}# }