Retrieve matched up neurons between the hemibrain and FAFB

Many neurons in the hemibrain data have been matched one to one with FAFB neurons (hemibrain->FAFB) and FAFB neurons have been matched with their hemibrain counterparts (FAFB->hemibrain). These matches have been done by the Flyconnectome Group at the University of Cambridge, and are recorded on a Google Sheet on our hemibrain Google Team Drive, to which you will need access through an authenticated account to view and use this function. Matches have three levels of 'quality', largely dependent on the degree of manual tracing for FAFB neurons - good (could be the same cell), medium (same cell_type) and poor (could be the same or similar cell_type).

hemibrain_matches(
  priority = c("FAFB", "hemibrain"),
  selected_file = options()$hemibrainr_matching_gsheet
)

lm_matches(
  priority = c("hemibrain", "lm"),
  selected_file = options()$hemibrainr_matching_gsheet
)

Arguments

priority
whether to use FAFB->hemibrain_matches (FAFB) or hemibrain->FAFB matches (hemibrain) in order to ascribe cell_type names to FAFB neurons. In both cases, cell_type names are attached to hemibrain bodyids, and propagated to their FAFB matches.

selected_file
the Google Sheet database to read and write from. For now, defaults to this Google Sheet.

Value

a data.frame which includes neuron's ID (either its CATMAID skeleton ID or neuprint body ID), the data set from which it comes, its putative cell_type and connectivity type, and its match in the other dataset.

a data.frame where each row is a neuron, either from the hemibrain or FAFB data sets. Each row gives you its matching neuron in the other data set. These matches have been manually assigned using fafb_matching, hemibrain_matching and LR_matching. If you use this information make sure you credit it appropriately. Contact us if unsure:

"cell_type" - the neuPrint designated 'type' for the neuron. If dataset is not "hemibrain", then this is based on the hemibrain match.

"cell" - the unique cell, which is just cell_type#number.

"cellBodyFiber" - the cell body fiber to which this neuron belongs

"ito_lee_hemilineage" - the hemilineage to which this neuron belongs. Seer hemibrain_hemilineages.

"match" - the ID of the manual match from the other data set. If dataset=="hemibrain" then this is a flywire_id that can be found in flywire_neurons.If "CATMAID" or "flywire" then it is a hemibrain body ID.

"quality" - the matcher makers qualitative assessment of how good this match is.

"fafb_hemisphere_match" - the flywire coordinates of a neuron on the opposite hemisphere, which a match maker has designated as this id's cognate.

"fafb_hemisphere_match.quality" - the quality of this match.

"LM.match" - indicates a light level neuron that is a match for id. This neuron will be in flycircuit_neurons() or other light level data.

"LM.match.quality" - the quality of this match.

"dataset" - the data set to which id belongs. Either "hemibrain", or one of the two FAFB repositories, "CATMAID" or "flywire".

"priority" - whether FAFB->hemibrain_matches ("FAFB") or hemibrain->FAFB matches ("hemibrain") were used in order to ascribe cell_type names to FAFB neurons. In both cases, cell_type names are attached to hemibrain bodyids, and propagated to their FAFB matches.

Details

Currently, the Google Sheet is set up with limited number of users, each of whom have been assigned a number of neurons to match up. In order to add yourself as a user, simply open this Google Sheet in your browser and add your initials to neurons of your choosing on the rightmost column 'Users'.

See also

Examples

# \donttest{
if (FALSE) {

# Get matches
matched = hemibrain_matches()

}# }