Match up neurons between the hemibrain and FAFB

Match up neurons between the hemibrain, FAFB and light level data and save the result using a Google Sheet on the hemibrain Google Team Drive operated by the flyconnectome group at the University of Cambridge. This function only considers lateral horn neurons. You must have access to the Team Drive in order to use this function. This function makes use of the Google Filestream application, which should be installed on your machine. Further, note that neurons are read from the FAFB CATMAID project, and you must have login details for this project recorded in your .Renviron for this function to work.

hemibrain_matching(
  ids = NULL,
  hemibrain.nblast = NULL,
  selected_file = "1OSlDtnR3B1LiB5cwI5x5Ql6LkZd8JOS5bBr-HTi0pOw",
  batch_size = 10,
  db = hemibrain_neurons(),
  match.type = c("FAFB", "LM"),
  query = NULL,
  overwrite = FALSE
)

lm_matching(
  ids = NULL,
  hemibrain.nblast = NULL,
  selected_file = "1OSlDtnR3B1LiB5cwI5x5Ql6LkZd8JOS5bBr-HTi0pOw",
  batch_size = 50,
  db = hemibrain_neurons(),
  query = NULL,
  overwrite = FALSE
)

fafb_matching(
  ids = NULL,
  hemibrain.nblast = NULL,
  selected_file = "1OSlDtnR3B1LiB5cwI5x5Ql6LkZd8JOS5bBr-HTi0pOw",
  batch_size = 50,
  db = hemibrain_neurons(),
  query = NULL,
  overwrite = FALSE
)

Arguments

ids
body IDs for hemibrain neurons present in the Google Sheet, for which the user will attempt to make a match if one has not been made already. Else, LM neuron IDs in the tab 'lm' when using lm_matching.

hemibrain.nblast
a FAFB (rows) - hemibrain (columns) normalised NBLAST matrix. By default this is read from the flyconnectome Team Drive.

selected_file
the Google Sheet database to read and write from. For now, defaults to a Google Sheet for lateral horn neurons. No other databases have been prepared.

batch_size
the number of FAFB top matches to read from CATMAID in one go.

db
Either a neuronlist or the name of a character vector naming a neuronlist. Defaults to the value of hemibrain_neurons().

match.type
whether to match up FAFB skeletons ("EM") or light level skeletons ("LM") from Dolan et al. and Frechter et al. 2019 (eLife), stored in the package lhns as most.lhns.

query
a neuronlist of light level neurons to match against. Should correspond to the given NBLAST matrix. Defaults to reading a transformed most.lhns from the Hemibrain Google Team Drive.

overwrite
logical, whether or not to overwrite matches already made.

Details

Currently, the Google Sheet is set up with limited number of users, each of whom have been assigned a number of neurons to match up. In order to add yourself as a user, simply open this Google Sheet in your browser and add your initials to neurons of your choosing on the rightmost column 'Users'. Once a match is recorded, the user selects a quality for that match. There can be no match (n), a poor match (p) an okay match (o) or an exact match (e). As a rule of thumb, a poor match could be a neuron from a very similar same cell type or a highly untraced neuron that may be the correct cell type. An okay match should be a neuron that looks to be from the same morphological cell type but there may be some discrepancies in its arbour. An exact match is a neuron that corresponds well between FAFB and the hemibrain data.

See also

Examples

# \donttest{
if (FALSE) {

# install package to bridge neurons between FAFB14 and hemibrain space
if (!requireNamespace("remotes")) install.packages("remotes")
remotes::install_github('natverse/nat.jrcbrains')
nat.jrcbrains::download_saalfeldlab_registrations()

# Load precomputed NBLAST from the flyconnectome Team Drive
load(file.path("/Volumes/GoogleDrive/Shared drives/flyconnectome/",
  "fafbpipeline/fib.fafb.crossnblast.twigs5.mean.compress.rda"))
# Sometimes the filenames are changed by drive
# load(file.path("/Volumes/GoogleDrive/Shared drives/flyconnectome",
# "fafbpipeline/fib.fafb.crossnblast.twigs5.mean.compress (1).rda"))

# Match!
hemibrain_matching(hemibrain.nblast = fib.fafb.crossnblast.twigs5.mean.compress)
}# }