Read neurons from the hemibrain connectome project

Read neurons from the hemibrain connectome project. This function uses the package neuprintr to read neurons. It then uses hemibrain_flow_centrality to re-root and split neurons into putative axons and dendrites. Optionally, it may also convert neurons from their raw voxel space (as they are stored in neuPrint) to microns.

hemibrain_read_neurons(
  x = NULL,
  savedir = TRUE,
  local = FALSE,
  microns = TRUE,
  remove.bad.synapses = TRUE,
  clean = TRUE,
  ...
)

hemibrain_neurons(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

Arguments

x
a vector of bodyids that can be read from 'https://neuprint.janelia.org/'.

savedir
directory in which nat::neuronlistfh object as been saved using hemibrain_download_neurons. This acts as a database, from which hemibrain neurons can be read. If TRUE your default save directory is used, which is stored as: options()$hemibrain_data

local
logical, whether to try to read locally saved neuronts (by detault at: options()$hemibrain_data) or neurons from Google Drive (options()$Gdrive_hemibrain_data).

microns
convert dimensions from raw voxels into microns (template brain: JRCFIB2018F, else JRCFIB2018Fraw).

remove.bad.synapses
whether or not to run hemibrain_remove_bad_synapses on neurons pulled from neuPrint.

clean
whether or not to set synapse-less branches to Label = 0.

...
methods passed to neuprintr::neuprint_read_neurons, hemibrain_remove_bad_synapses and hemibrain_flow_centrality

neuron.split
read saved neurons spit in which way? Folder names indicative of arguments passed to flow_centrality.

Value

the neuron or neuron list object inputted, with centripetal flow centrality information added to neuron$d and a segregation index score. The neuron$d$Label now gives the compartment, where axon is Label = 2, dendrite Label = 3, primary dendrite Label = 9 and primary neurite Label = 7. Soma is Label = 1.

See also

Examples

# \donttest{
if (FALSE) {

# Choose neurons
## In this case some antennal lobe local neurons
al.local.neurons = c("1702323386", "2068966051", "2069311379", "1702305987", "5812996027",
"1702336197", "1793744512", "1976565858", "2007578510", "2101339904",
"5813003258", "2069647778", "1947192569", "1883788812", "1916485259",
"1887177026", "2101348562", "2132375072", "2256863785", "5813002313",
"5813054716", "5813018847", "5813055448", "1763037543", "2101391269",
"1794037618", "5813018729", "2013333009")

# Get neurons
neurons = hemibrain_read_neurons(al.local.neurons)

# Plot the split to check it
nat::nopen3d()
nlscan_split(neurons, WithConnectors = TRUE)
}# }